feat!: define an Aggregate or Projection to be enriched if any events were applied to it.
Previously, isDefined/isEnriched is defined by whether the Projection were enriched
with events from a system that is not empty. Therefore,
the only way for isDefined to be false is if the system is empty.

However, given that the Projection can also filter the system events
based on stream names and/or event types, the default
definition for isDefined seems to be less useful.

Likewise, an Aggregate would never be undefined after enrichment because
it was defined by the last version of the event stream. And the only way
for the last version to be "undefined" is when the stream does not exist.
However, the enrichment would throw a "stream does not exist" exception
(correctly) in this case. Therefore, it is impossible for an Aggregate to be
undefined.

This change reconciles the differences between Aggregate and Projection and
achieves a simpler definition of "isDefined". They are defined when any
events were applied to them. As such, the "isDefined" property is renamed to
"isEnriched" to increase clarity.